backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
python-dotenv
httpx
sqlalchemy     
psycopg2-binary 
sqlalchemy-utils